+++ sdk/core/azure-core/CMakeLists.txt
@@ -47,9 +47,7 @@ set(
     ${WIN_TRANSPORT_ADAPTER_INC}
     inc/azure/core/credentials/credentials.hpp
     inc/azure/core/credentials/token_credential_options.hpp
-    inc/azure/core/cryptography/hash_base.hpp
     inc/azure/core/cryptography/hash.hpp
-    inc/azure/core/cryptography/md5.hpp
     inc/azure/core/diagnostics/logger.hpp
     inc/azure/core/http/http_status_code.hpp
     inc/azure/core/http/http.hpp
